+++ Utilities/Utility.cs
@@ -575,9 +575,9 @@ namespace SeedFinding
             r.Next();
             return r.NextDouble() < 0.1;
         }
-        public static Random CreateDaySaveRandom(int days, uint gameId, double seedA = 0.0, double seedB = 0.0, double seedC = 0.0)
+        public static Random CreateDaySaveRandom(int days, long gameId, double seedA = 0.0, double seedB = 0.0, double seedC = 0.0)
         {
-            return Utility.CreateRandom(days, gameId / 2uL, seedA, seedB, seedC);
+            return Utility.CreateRandom(days, gameId / 2, seedA, seedB, seedC);
         }
 
         public static Random CreateRandom(double seedA, double seedB = 0.0, double seedC = 0.0, double seedD = 0.0, double seedE = 0.0)
@@ -629,7 +629,7 @@ namespace SeedFinding
             return list[random.Next(list.Count)];
         }
 
-        public static bool isGreenRainDay(int day, uint gameId)
+        public static bool isGreenRainDay(int day, long gameId)
         {
             Season season = getSeasonFromDay(day);
             if (season == Season.Summer)
@@ -640,7 +640,7 @@ namespace SeedFinding
             }
             return false;
         }
-        public static int getRandomItemFromSeason(Season season, int randomSeedAddition, bool forQuest, int all_unlocked_cooking_recipes, uint gameId, int daysPlayed, bool changeDaily = true, bool hasFurnace = false, bool hasDesert = false, int mines = 0)
+        public static int getRandomItemFromSeason(Season season, int randomSeedAddition, bool forQuest, int all_unlocked_cooking_recipes, long gameId, int daysPlayed, bool changeDaily = true, bool hasFurnace = false, bool hasDesert = false, int mines = 0)
         {
             int dayOfMonth = (daysPlayed - 1) % 28 + 1;
             int year = (daysPlayed - 1) / (28 * 4) + 1;
